requirements.txt
apache-airflow==2.10.5
fastapi==0.115.11
happybase==1.2.0
ipython==8.12.3
kafka_python==2.1.4
langchain==0.3.20
langchain_community==0.3.19
langchain_core==0.3.41
langchain_openai==0.3.7
langserve==0.3.1
matplotlib==3.10.0
minio==7.2.15
networkx==3.4.2
neo4j==5.27.0
nltk==3.9.1
numpy==1.26.4
overrides==7.4.0
pandas==2.2.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.12.0
scikit_learn==1.6.1
scikit_surprise==1.1.4
sentence_transformers==3.3.1
streamlit==1.42.1
surprise==0.1
tensorboard==2.18.0
torch_geometric==2.6.1
tqdm>=4.66.6
uvicorn==0.34.0
yfiles_jupyter_graphs==1.9.0
